requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
alembic==1.13.1
pymysql==1.1.1
python-dotenv==1.0.1
pydantic-settings==2.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
librosa==0.10.2
scikit-learn==1.5.0
numpy==1.26.4
filelock==3.14.0
mutagen==1.47.0
pytest==8.2.0
httpx==0.27.0
openai-whisper>=20231117
ffmpeg-python>=0.2.0